Refactor ram transforms into separate files
