+++ src/main.cpp
@@ -66,8 +66,21 @@
 #include "ram/RamNode.h"
 #include "ram/RamProgram.h"
 #include "ram/RamTranslationUnit.h"
+#include "ram/transform/ChoiceConversion.h"
+#include "ram/transform/CollapseFilters.h"
+#include "ram/transform/EliminateDuplicates.h"
+#include "ram/transform/ExpandFilter.h"
+#include "ram/transform/HoistAggregate.h"
+#include "ram/transform/HoistConditions.h"
+#include "ram/transform/IfConversion.h"
+#include "ram/transform/IndexedInequality.h"
+#include "ram/transform/MakeIndex.h"
+#include "ram/transform/Parallel.h"
 #include "ram/transform/RamTransformer.h"
-#include "ram/transform/RamTransforms.h"
+#include "ram/transform/ReorderConditions.h"
+#include "ram/transform/ReorderFilterBreak.h"
+#include "ram/transform/ReportIndex.h"
+#include "ram/transform/TupleId.h"
 #include "synthesiser/Synthesiser.h"
 #include "utility/FileUtil.h"
 #include "utility/StreamUtil.h"
