+++ src/ram/transform/IfConversion.cpp
@@ -1,0 +1,99 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2018, The Souffle Developers. All rights reserved
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file IfConversion.cpp
+ *
+ ***********************************************************************/
+
+#include "ram/transform/IfConversion.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamVisitor.h"
+#include "utility/MiscUtil.h"
+#include <algorithm>
+#include <cstddef>
+#include <functional>
+#include <utility>
+#include <vector>
+
+namespace souffle {
+
+std::unique_ptr<RamOperation> IfConversionTransformer::rewriteIndexScan(const RamIndexScan* indexScan) {
+    // check whether tuple is used in subsequent operations
+    bool tupleNotUsed = true;
+    visitDepthFirst(*indexScan, [&](const RamTupleElement& element) {
+        if (element.getTupleId() == indexScan->getTupleId()) {
+            tupleNotUsed = false;
+        }
+    });
+
+    // if not used, transform the IndexScan operation to an existence check
+    if (tupleNotUsed) {
+        // replace IndexScan with an Filter/Existence check
+        std::vector<std::unique_ptr<RamExpression>> newValues;
+
+        size_t arity = indexScan->getRelation().getArity();
+        for (size_t i = 0; i < arity; ++i) {
+            if (*(indexScan->getRangePattern().first[i]) != *(indexScan->getRangePattern().second[i])) {
+                fatal("Inequal upper and lower bounds not supported while rewriting index scan in "
+                      "IfConversionTransformer");
+            }
+        }
+
+        for (auto& cur : indexScan->getRangePattern().second) {
+            RamExpression* val = nullptr;
+            if (cur != nullptr) {
+                val = cur->clone();
+            }
+            newValues.emplace_back(val);
+        }
+
+        // check if there is a break statement nested in the Scan - if so, remove it
+        RamOperation* newOp;
+        if (const auto* breakOp = dynamic_cast<const RamBreak*>(&indexScan->getOperation())) {
+            newOp = breakOp->getOperation().clone();
+        } else {
+            newOp = indexScan->getOperation().clone();
+        }
+
+        return std::make_unique<RamFilter>(
+                std::make_unique<RamExistenceCheck>(
+                        std::make_unique<RamRelationReference>(&indexScan->getRelation()),
+                        std::move(newValues)),
+                std::unique_ptr<RamOperation>(newOp), indexScan->getProfileText());
+    }
+    return nullptr;
+}
+
+bool IfConversionTransformer::convertIndexScans(RamProgram& program) {
+    bool changed = false;
+    visitDepthFirst(program, [&](const RamQuery& query) {
+        std::function<std::unique_ptr<RamNode>(std::unique_ptr<RamNode>)> scanRewriter =
+                [&](std::unique_ptr<RamNode> node) -> std::unique_ptr<RamNode> {
+            if (const RamIndexScan* scan = dynamic_cast<RamIndexScan*>(node.get())) {
+                if (std::unique_ptr<RamOperation> op = rewriteIndexScan(scan)) {
+                    changed = true;
+                    node = std::move(op);
+                }
+            }
+            node->apply(makeLambdaRamMapper(scanRewriter));
+            return node;
+        };
+        const_cast<RamQuery*>(&query)->apply(makeLambdaRamMapper(scanRewriter));
+    });
+    return changed;
+}
+
+}  // end of namespace souffle
